backend/functions/processing/requirements.txt
pillow==7.2.0
requests==2.24.0
google-cloud-storage==1.31.2
numpy==1.19.0
